Add sperated unittests
