+++ .travis.yml
@@ -19,7 +19,7 @@ install:
 #  - python manage.py migrate
 
 script:
-  - coverage run manage.py test polls
+  - coverage run manage.py tests polls
 
 after_script:
   - bash <(curl -s https://codecov.io/bash)
