+++ requirements.txt
@@ -1,4 +1,3 @@
-decouple
 coverage
 python-decouple
 django
